requirements.txt
transformers
--extra-index-url https://download.pytorch.org/whl/cu118

# Flask & Extensions
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Cors==4.0.0
Flask-WTF==1.2.1

# Database
PyMySQL==1.1.0
cryptography==42.0.0

# Authentication
bcrypt==4.1.2
PyJWT==2.8.0

# OCR + Image Processing
easyocr==1.7.1
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.26.4
scikit-image==0.21.0

# PyTorch GPU (RTX 3050 - CUDA 11.8) + Python 3.11 compatible
torch==2.1.2+cu118
torchvision==0.16.2+cu118
torchaudio==2.1.2+cu118

# Text Processing
pyspellchecker==0.7.2
unidecode==1.3.7

# TTS
gTTS==2.5.1
# Translation
googletrans==4.0.2

# Utilities
python-dotenv==1.0.0
Werkzeug==3.0.1

sentencepiece

# Testing
pytest==8.0.0
hypothesis==6.92.0
